Chore/restructure server (#22)
* chore: api docs

* chore: split server src files to folders
