+++ server/src/action/MoveAction.cpp
@@ -1,5 +1,5 @@
 #include "MoveAction.hpp"
-#include "PlayerCharacter.hpp"
+#include "../entity/PlayerCharacter.hpp"
 
 #include <iostream>
