+++ server/src/entity/Entity.cpp
